Update config.yaml
hmm, 
not sure if it is going to work out.
The reason for this change is going to be that i want to push specific topics into specific tables anyways, sooo it is convenient to do it right away. 
the other options are a) dump anything into 1 table and sort it on postgres level or even later (delete old stuff to keep the dump small)
b) do it like this
c) multiple instances with 1 connection each for lots of topics, 
d) ... forgot that one

present-me assumes this solution is the most convenient one because all the topics are in one config file and the comments might help a lot if configurations have to be redone once in a while.